Move common validations code to seperate object
