+++ engine/flink/management/sample/src/main/resources/reference.conf
@@ -5,8 +5,7 @@ additionalPropertiesConfig {
     }
     validators: [
       {
-        type: "MandatoryParameterValidator",
-
+        type: "MandatoryParameterValidator"
       }
     ]
     label: "Environment"
